scraper/DC_Events_Scraper/scraper/requirements.txt
requests
datetime
python-dateutil
boto3
bs4

